refactor: update namespaces and using directives
