+++ LotCoMClient/Models/Options/DataTablePageOptions.cs
@@ -1,0 +1,1 @@
+namespace LotCoMClient.Models.Options;
